# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: |
    API untuk sistem marketplace dengan fitur:
    - Manajemen produk
    - Sistem reseller dengan komisi
    - Tracking click dan leads
    - Manajemen payout
    - Integrasi MinIO untuk upload file
  version: 1.0.0
  contact:
    name: API Support
    email: support@marketplace.com
servers:
  - url: http://localhost:8080/api
    description: Development Server
  - url: https://api.marketplace.com/api
    description: Production Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token yang didapat dari endpoint login
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
        status:
          type: string
          enum: [AVAILABLE, SOLD, INACTIVE]
        location_name:
          type: string
        location_province:
          type: string
        location_regency:
          type: string
        location_district:
          type: string
        location_village:
          type: string
        province_code:
          type: string
        regency_code:
          type: string
        district_code:
          type: string
        village_code:
          type: string
        commission_amount:
          type: integer
        contact_phone:
          type: string
        contact_wa:
          type: string
        thumbnail_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Payout:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reseller_id:
          type: string
          format: uuid
        amount:
          type: integer
        proof_object_key:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
paths:
...
  /products:
    get:
      tags:
        - Public
      summary: List produk dengan filter
      description: Mendapatkan daftar produk dengan opsi filter dan pagination
      parameters:
        - name: category
          in: query
          description: Filter berdasarkan kategori
          schema:
            type: string
        - name: search
          in: query
          description: Pencarian berdasarkan judul atau deskripsi
          schema:
            type: string
        - name: province
          in: query
          description: Filter berdasarkan kode provinsi
          schema:
            type: string
        - name: limit
          in: query
          description: Jumlah data per halaman (default 20)
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset untuk pagination (default 0)
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                    example: "PRODUCTS_RETRIEVED"
                  message_data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      total:
                        type: integer
...
  /admin/products:
    get:
      tags:
        - Admin
      summary: List all products (admin)
      description: Admin mendapatkan semua produk dengan pagination
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Success

    post:
      tags:
        - Admin
      summary: Create new product
      description: Admin membuat produk baru
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - price
                - category
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: integer
                category:
                  type: string
                location_name:
                  type: string
                location_province:
                  type: string
                location_regency:
                  type: string
                location_district:
                  type: string
                location_village:
                  type: string
                province_code:
                  type: string
                regency_code:
                  type: string
                district_code:
                  type: string
                village_code:
                  type: string
                commission_amount:
                  type: integer
                contact_phone:
                  type: string
                contact_wa:
                  type: string
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                    example: "PRODUCT_CREATED"
                  message_data:
                    $ref: '#/components/schemas/Product'
...
  /admin/payouts:
    get:
      tags:
        - Admin
      summary: List all payouts
      description: Admin mendapatkan semua riwayat payout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                  message_data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payout'

    post:
      tags:
        - Admin
      summary: Create payout
      description: Admin membuat payout untuk reseller
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - reseller_id
                - amount
              properties:
                reseller_id:
                  type: string
                  format: uuid
                amount:
                  type: integer
                  description: Jumlah payout dalam rupiah
                notes:
                  type: string
                  description: Catatan tambahan
                proof:
                  type: string
                  format: binary
                  description: Bukti transfer (gambar)
      responses:
        '201':
          description: Payout created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_call_id:
                    type: string
                  message_action:
                    type: string
                    example: "PAYOUT_CREATED"
                  message_data:
                    $ref: '#/components/schemas/Payout'

  # ==================== CLIENT/RESELLER ENDPOINTS ====================
...